refactor(workflow-runtime): extract submission tests via #[path]
Phase 0 PR-2 of the architecture decoupling roadmap (#1082 was PR-1).
Same pattern: move workflow_runtime_submission.rs's `#[cfg(test)] mod
tests` block into a separate file via `#[cfg(test)] #[path = "..."]
mod tests;`.

Net effect:
  workflow_runtime_submission.rs:        2280 -> 1254 lines (-1026)
  workflow_runtime_submission_tests.rs:   NEW   1017 lines

Zero logic changes. The new file uses `use super::*;` so it sees the
submission module unchanged. No public API surface change.

Verification:
- cargo test -p harness-server --lib -- workflow_runtime_submission:
    17/17 passed (including http::tests::create_task_with_* which
    exercise the submission path through the HTTP handler)
- RUSTFLAGS=-Dwarnings cargo check --workspace --all-targets: clean
- cargo fmt --all -- --check: clean

submission.rs is still over the U-16 800-line ceiling (1254 lines
remaining, mostly persistence + decision builders) but no longer the
second-largest module. Future Phase 1 cohesive splits can target the
remaining production code.
